% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_birth_data.R
\name{create_birth_data_raw}
\alias{create_birth_data_raw}
\title{Create dummy birth data}
\usage{
create_birth_data_raw(n = 10000, seed = 249)
}
\arguments{
\item{n}{Number of rows to generate. Default is 1000.}

\item{seed}{Numerical value of seed for reproducibility. Default is 249.}
}
\value{
Data frame.
}
\description{
Some messy data is added to demonstrate processing.
}
