% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{derive_age}
\alias{derive_age}
\title{Derive age column in data frame}
\usage{
derive_age(data, age_col, dob_col, end_date_col)
}
\arguments{
\item{data}{Data frame.}

\item{age_col}{The unquoted name of the new column to store the calculated
age.}

\item{dob_col}{The unquoted name of the Date type column containing the date
of birth values.}

\item{end_date_col}{The unquoted name of the Date type column containing the
end date of calculation.}
}
\value{
Data frame with added age column.
}
\description{
Uses a date of birth column and a column for the date to
calculate to. Dynamically selects what to name the column. Errors if the
columns provided are not Date type.
}
