% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{derive_live_or_still}
\alias{derive_live_or_still}
\title{Derive live or still birth from variable codes}
\usage{
derive_live_or_still(data, fcts)
}
\arguments{
\item{data}{Data frame with \code{birth_type} column.}

\item{fcts}{List of character vectors with the \code{live_or_still} factor levels.}
}
\value{
Original data frame with recoded variable.
}
\description{
The \code{birth_type} column is replaced with \code{live_or_still} factor
with two levels: "Live birth" and "Stillbirth".
}
