% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{derive_timeliness}
\alias{derive_timeliness}
\title{Derive timeliness (registration delay)}
\usage{
derive_timeliness(
  data,
  date_event_col,
  date_reg_col,
  threshold_late = 30,
  threshold_delayed = 365
)
}
\arguments{
\item{data}{Data frame.}

\item{date_event_col}{The unquoted name of the column containing the date of
event.}

\item{date_reg_col}{The unquoted name of the column containing the date of
registration of the event.}

\item{threshold_late}{Numeric value of when a registration goes from current
to late. Default is 30. The late threshold is not inclusive for "current"
(e.g. a 30 day delay with a late threshold of 30 would be late).}

\item{threshold_delayed}{Numeric value of when a registration goes from late
to delayed Default is 365. The delayed threshold is not inclusive for "late"
(e.g. a 365 day delay with a delayed threshold of 365 would be delayed).}
}
\value{
Data frame with data1c (days between event and registration) and
data2c (category of timeliness) columns.
}
\description{
The time between event and registration (registration delay) is
a valuable metric for assessing if the civil registration process is working.
The delay can be classified whether the delay is within an acceptable
threshold, late or delayed. These thresholds can be based on UN guidelines or
laws within a country.
}
\details{
The categories are: current, late and then delayed.
}
