% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{derive_var}
\alias{derive_var}
\title{Convert to factor with levels, no NA}
\usage{
derive_var(data, col_str, fcts_vector)
}
\arguments{
\item{data}{Data frame containing variable to change.}

\item{col_str}{String of the column to change.}

\item{fcts_vector}{Character vector of the variable to use as factor levels.}
}
\value{
Original data frame with changed variable
}
\description{
Convert NA to "Not stated" and change the column to a factor
with defined factor levels.
}
