% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{derive_year}
\alias{derive_year}
\title{Derive year}
\usage{
derive_year(data, date_col, year_col = year)
}
\arguments{
\item{data}{Data frame.}

\item{date_col}{The unquoted name of the column containing the date to
extract the year from. The date must Date type in format "YYYY-MM-DD".}

\item{year_col}{The unquoted name of the new column to store the year.
Default is year.}
}
\value{
Data frame with new column for year.
}
\description{
Derive year
}
