% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factorise_cols}
\alias{factorise_cols}
\title{Convert selected data.frame columns to factors using a named list of levels}
\usage{
factorise_cols(df, level_list, cols = names(level_list), strict = TRUE)
}
\arguments{
\item{df}{A data.frame (or tibble).}

\item{level_list}{A named list. Each name must match a column in df.
Each value must be a character vector of desired levels.}

\item{cols}{Optional character vector of column names to convert.
Defaults to names(level_list).}

\item{strict}{Logical; if TRUE, error when df has values not found in
provided levels. If FALSE, those values become NA with a warning.}
}
\value{
The modified data.frame with selected columns turned into factors.
}
\description{
Convert selected data.frame columns to factors using a named list of levels
}
