% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_table_functions.R
\name{general_count_by_var}
\alias{general_count_by_var}
\title{General function to count by chosen aggregation columns}
\usage{
general_count_by_var(
  data,
  agg_cols,
  agg_cols_pub_names,
  agg_count_colname = "Counts of births",
  count_or_sum = "count",
  sum_col = NULL,
  fcts,
  pre_agg_levels_to_keep = list(),
  post_agg_levels_to_keep = list()
)
}
\arguments{
\item{data}{Processed data frame (including factorising aggregation
variables).}

\item{agg_cols}{Vector of characters of input data frame variable names of
variables to aggregate by.}

\item{agg_cols_pub_names}{Vector of characters of intended names to rename
the input aggregation variables to (must be same length as
\code{agg_cols_pub_names}). Should be publication ready.}

\item{agg_count_colname}{Symbol (unquoted or ``) to name the aggregated
count column. Default is \verb{Counts of births}.}

\item{fcts}{List object of character vectors of factor levels within each of
the aggregation variables from \code{config.R}.}

\item{pre_agg_levels_to_keep}{List object of character vectors to filter the
input dataset before aggregation.
Names of list items must relate to the input column names (\code{agg_cols}).
Items in the list are the factor levels to keep.
Default is no filers applied. See \code{filter_factor_levels} for more information
on the filtering.}

\item{post_agg_levels_to_keep}{List object of character vectors to filter the
input dataset after aggregation.
Names of list items must relate to the input column names (\code{agg_cols}).
Items in the list are the factor levels to keep.
Default is no filers applied. See \code{filter_factor_levels} for more information
on the filtering.}
}
\value{
Aggregated counts of chosen variables and breakdowns.
}
\description{
General function to count by chosen aggregation columns
}
