% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{group_age}
\alias{group_age}
\title{Group age column as specified}
\usage{
group_age(data, age_col, new_col, start_brk, end_brk, brk_step, col_factor)
}
\arguments{
\item{data}{Data frame.}

\item{age_col}{Quoted name of age column.}

\item{new_col}{Quoted name of new column for derived groups.}

\item{start_brk}{Integer for where granular age groups begin (anything less
grouped together).}

\item{end_brk}{Integer for where granular age groups end (anything greater
lumped into one group).}

\item{brk_step}{Integer for size of age groups (e.g. if 5, then groups of
20 - 24, 25 - 29, ...).}

\item{col_factor}{Vector containing the order of values.}
}
\value{
Data frame with new age group column.
}
\description{
Group age column as specified
}
