% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_by_summarise_sum}
\alias{group_by_summarise_sum}
\title{Sum values by grouping variables}
\usage{
group_by_summarise_sum(data, grouping_cols, sum_col)
}
\arguments{
\item{data}{A data frame containing at least the \code{grouping_cols} and
\code{sum_col}.}

\item{grouping_cols}{Vector of character strings of column names to group by.}

\item{sum_col}{A character string of the numeric column name to sum.}
}
\value{
A data frame with one row per unique combination of grouping
variables and a column containing the summed values.
}
\description{
Groups by one or more chosen variables/columns
(\code{grouping_cols}), calculates the sum of a chosen numeric variable/column
(\code{sum_col}) in each group. All combinations of factor variable levels will be
returned, even with no instances (i.e. counts of 0).
}
\examples{
df <- data.frame(
  region = c("A", "A", "B"),
  sex = c("Male", "Female", "Male"),
  value = c(10, 20, 30)
)

group_by_summarise_sum(df, c("region", "sex"), "value")

}
