% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{make_na_not_stated}
\alias{make_na_not_stated}
\title{Convert a missing/NA into "Not stated"}
\usage{
make_na_not_stated(data, colname)
}
\arguments{
\item{data}{Data frame with the column to convert.}

\item{colname}{String of column name to change.}
}
\value{
Original data frame with the column converted.
}
\description{
Processing sub-function that cleans data by converting NA into
"Not stated" or keeps the original value, and makes the column character).
Used in other processing functions.
}
