% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_birth_data.R
\name{process_dummy_birth_data}
\alias{process_dummy_birth_data}
\title{Process dummy birth data - ADD INFO WHEN FINISHED.}
\usage{
process_dummy_birth_data(data, col_factor)
}
\arguments{
\item{data}{Data frame of record-level birth data.}
}
\value{
Processed data frame.
}
\description{
Process dummy birth data - ADD INFO WHEN FINISHED.
}
