% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_asfr_usual_res.R
\name{process_pop_for_asfr}
\alias{process_pop_for_asfr}
\title{Process population data for Age-Specific Fertility Rate calculation}
\usage{
process_pop_for_asfr(pop, fcts, pop_sex = "Female", fert_range = 15:44)
}
\arguments{
\item{pop}{A data frame containing at least the columns:
\code{year}, \code{geog}, \code{age}, and \code{population}.}

\item{fcts}{A list containing lookup vectors or functions used for grouping,
specifically \code{mother_age_group} for mapping age groups.}

\item{pop_sex}{Character string containing either "Female", "Male" or
"All people". Default is "Female".}

\item{fert_range}{An integer vector giving the fertility age range.
Defaults to \code{15:44}.}
}
\value{
A data frame containing:
\itemize{
\item Population totals by year, geography, and 5‑year age group;
\item A row where all age groups are aggregated into \code{"All ages"};
\item A modified youngest group labelled \code{"Under 20"}.
}
}
\description{
This function filters population data to the fertility age range,
groups ages into 5-year bands, aggregates population counts,
and prepares age‑grouped totals for ASFR calculations.
}
