% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{rename_crvs_columns}
\alias{rename_crvs_columns}
\title{Rename columns in a data frame using a lookup table}
\usage{
rename_crvs_columns(
  df,
  lookup,
  old_names = input_variable_id,
  new_names = standardised_variable_name
)
}
\arguments{
\item{df}{A data frame with columns to be renamed.}

\item{lookup}{A data frame containing the lookup between old and new column
names.}

\item{old_names}{An unquoted column name in \code{lookup} containing the existing
column names in \code{df}. Defaults to \code{variable_id}.}

\item{new_names}{An unquoted column name in \code{lookup} containing the new
column names to apply. Defaults to \code{variable_name}.}
}
\value{
A data frame with renamed columns or original columns if no match in
lookup.
}
\description{
Renames columns in a data frame based on a lookup table that maps old column
names to new column names. Column names are supplied using tidy‑evaluation
(data masking), so should be provided without quote marks.
}
\details{
The function renames matching columns according to the lookup table.
Columns in \code{df} that do not appear in the lookup table are left unchanged.
}
