% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_multi_sum_str}
\alias{rename_multi_sum_str}
\title{Rename 'Sum' in multiple columns to informative string}
\usage{
rename_multi_sum_str(data, cols, vars, pub_colnames, fcts)
}
\arguments{
\item{data}{Data frame.}

\item{cols}{Vector of string column names to rename and change 'Sum' to
correct factor level.}

\item{vars}{Vector of strings of variables that correspond to the names of
the list of factor levels from the config and also match the \code{sum_to_string}
function.}

\item{pub_colnames}{Vector strings of new column names for the \code{vars}
columns.}
}
\value{
Formatted data frame.
}
\description{
Also renames the column names.
}
