% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_sum_str}
\alias{rename_sum_str}
\title{Rename 'Sum' in columns to informative string}
\usage{
rename_sum_str(data, col, var, pub_colname, fcts)
}
\arguments{
\item{data}{Data frame.}

\item{col}{Name of column to mutate as a string.}

\item{var}{Quoted string of name of variable that corresponds with the names
of the vectors of factor levels in the config and \code{sum_str}.}

\item{pub_colname}{Quoted column name for publishing.}

\item{fcts}{List of character vectors from config for the factor levels to
use.}
}
\value{
Data frame.
}
\description{
Rename 'Sum' in columns to informative string
}
