% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pops_data.R
\name{create_population_data}
\alias{create_population_data}
\title{Create wide dummy populations data - Not currently in use}
\usage{
create_population_data(seed = 249)
}
\arguments{
\item{seed}{Numerical value of seed for reproducibility. Default is 249.}
}
\value{
Data frame.
}
\description{
Create wide dummy populations data - Not currently in use
}
