% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_table_functions.R
\name{general_count_by_var}
\alias{general_count_by_var}
\title{General function to count by chosen aggregation columns}
\usage{
general_count_by_var(
  data,
  agg_cols,
  agg_cols_pub_names,
  agg_count_colname = "Counts of births",
  count_or_sum = "count",
  sum_col = NULL,
  fcts,
  pre_agg_levels_to_keep = list(),
  post_agg_levels_to_keep = list()
)
}
\arguments{
\item{data}{Processed data frame (including factorising aggregation
variables).}

\item{agg_cols}{Vector of characters of input data frame variable names of
variables to aggregate by.}

\item{agg_cols_pub_names}{Vector of characters of intended names to rename
the input aggregation variables to (must be same length as
\code{agg_cols_pub_names}). Should be publication ready.}

\item{agg_count_colname}{Name of the intended output column containing counts
or sums. Default is \code{"Counts of births"}.}

\item{count_or_sum}{String. Whether to \code{"count"} rows or \code{"sum"} during
aggregation. Default is \code{"count"}.}

\item{sum_col}{String. Column to sum by when \code{count_or_sum = "sum"}.
Ignored otherwise (default is \code{NULL}.}

\item{fcts}{List object of character vectors of factor levels within each of
the aggregation variables from \code{config.R}.}

\item{pre_agg_levels_to_keep}{List object of character vectors to filter the
input dataset before aggregation.
Names of list items must relate to the input column names (\code{agg_cols}).
Items in the list are the factor levels to keep.
Default is no filers applied. See \code{filter_factor_levels} for more information
on the filtering.}

\item{post_agg_levels_to_keep}{List object of character vectors to filter the
input dataset after aggregation.
Names of list items must relate to the input column names (\code{agg_cols}).
Items in the list are the factor levels to keep.
Default is no filers applied. See \code{filter_factor_levels} for more information
on the filtering.}
}
\value{
A data frame grouped by the selected variables, with counts or sums and
cleaned, publication-ready labels.
}
\description{
This function groups data by one or more variables and calculates either
counts or sums. It also formats the output so it is ready for reporting,
including replacing \code{"Sum"} labels and applying readable column names.

You can choose to drop chosen factor levels before and also after
aggregation.
}
\details{
The function:
\itemize{
\item Checks required columns exist
\item Filters factor levels (optional, before and after aggregation)
\item Calculates counts or sums
\item Replaces \code{"Sum"} with meaningful labels
\item Sorts and cleans the final output
}
}
\examples{
general_count_by_var(
  data = births,
  agg_cols = c("sex", "year"),
  agg_cols_pub_names = c("Sex", "Year"),
  fcts = sum_to_string()
)

}
