% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_levels.R
\name{sum_to_string}
\alias{sum_to_string}
\title{List of strings to replace "Sum" with in columns}
\usage{
sum_to_string()
}
\value{
List object of strings to use for publication-ready totals.
}
\description{
The process of calculating totals (\code{rename_multi_sum_str},
\code{rename_sum_str}) for all combinations of multiple variables creates a factor
level of "Sum" which should be re-coded to the true factor level.
The name of the string should match the name of the vector from the config.
}
