% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_count_vars}
\alias{summarise_count_vars}
\title{Create cross-tabulated counts or sums with totals added}
\usage{
summarise_count_vars(
  data,
  pub_name_col,
  grouping_cols,
  count_or_sum = "count",
  sum_col
)
}
\arguments{
\item{data}{Processed dataframe with no NA in grouping columns (should be
"Not stated" or similar).}

\item{grouping_cols}{Vector of strings of column names to be used when
grouping, summarising and counting.}

\item{count_or_sum}{String of "count" or "sum" to define if the chosen column
should count by rows or sum. Default is "count".}

\item{sum_col}{What is the column to be summed if \code{count_or_sum} is "sum".
Should be supplied unquoted. Default is NULL.}

\item{col}{Unquoted name to rename the count column with.}
}
\value{
Dataframe.
}
\description{
Aggregates a dataset by specified grouping variables
(\code{grouping_cols}), producing either counts or sums,
and returns a cross-tabulation (all combinations of the variables) with the
totals/sums (marginal totals) calculated.
}
